# PERFORMANCE: LIR Pipeline Execution