--- a/logstash-core/src/main/java/org/logstash/config/ir/PluginDefinition.java
+++ b/logstash-core/src/main/java/org/logstash/config/ir/PluginDefinition.java
@@ -11,7 +11,7 @@ import org.logstash.common.SourceWithMetadata;
 /**
  * Created by andrewvc on 9/20/16.
  */
-public class PluginDefinition implements SourceComponent, HashableWithSource {
+public final class PluginDefinition implements SourceComponent, HashableWithSource {
 
     @Override
     public String hashSource() {