… logstash-core/lib/logstash/java_pipeline.rb
… logstash-core/locales/en.yml
… logstash-core/spec/conditionals_spec.rb
… logstash-core/spec/logstash/filters/base_spec.rb
… logstash-core/spec/logstash/java_filter_delegator_spec.rb
… logstash-core/spec/logstash/java_pipeline_spec.rb
… logstash-core/spec/support/helpers.rb
… logstash-core/spec/support/pipeline/pipeline_helpers.rb
… logstash-core/src/main/java/org/logstash/config/ir/DSL.java
… logstash-core/src/main/java/org/logstash/config/ir/compiler/Dataset.java
… logstash-core/src/main/java/org/logstash/config/ir/compiler/EventCondition.java
… logstash-core/src/main/java/org/logstash/config/ir/compiler/RubyIntegration.java
… logstash-core/src/main/java/org/logstash/config/ir/expression/Expression.java
… logstash-core/src/main/java/org/logstash/config/ir/expression/RegexValueExpression.java
… logstash-core/src/main/java/org/logstash/config/ir/expression/ValueExpression.java
… logstash-core/src/main/java/org/logstash/config/ir/graph/BooleanEdge.java
… logstash-core/src/main/java/org/logstash/config/ir/imperative/PluginStatement.java
… logstash-core/src/test/java/org/logstash/common/io/DeadLetterQueueReaderTest.java