@original-brownbear 2017-09-21
@ph @andrewvc let's continue here? The other one became unusable laggy in browsers imo, I'd close it if you're cool with it? (I'll add a proper description here soon then :))