@original-brownbear 2017-09-22
@andrewvc @ph Updated performance numbers :)

### Master

```sh
➜  logstash git:(master) time cat ~/Downloads/apache_access_logs ~/Downloads/apache_access_logs | bin/logstash -b 128 -w 4 -f ~/tmp/logstash.cfg | pv | wc -c
13.2MiB 0:07:01 [32.0KiB/s] [                                                   <=>                                                                                                                                                                                                                                                                             ]
 13802035
cat ~/Downloads/apache_access_logs ~/Downloads/apache_access_logs  0.17s user 2.91s system 0% cpu 7:00.52 total
bin/logstash -b 128 -w 4 -f ~/tmp/logstash.cfg  1755.94s user 25.90s system 422% cpu 7:01.41 total
pv  3.24s user 19.75s system 5% cpu 7:01.40 total
wc -c  0.42s user 3.34s system 0% cpu 7:01.40 total
```

### This

```sh
➜  logstash git:(lir-if-else-squashed) time cat ~/Downloads/apache_access_logs ~/Downloads/apache_access_logs | bin/logstash -b 128 -w 4 -f ~/tmp/logstash.cfg | pv | wc -c
13.2MiB 0:04:37 [48.5KiB/s] [                                                                                                                                                                                                                         <=>                                                                                                       ]
 13801638
cat ~/Downloads/apache_access_logs ~/Downloads/apache_access_logs  0.12s user 2.25s system 0% cpu 4:36.88 total
bin/logstash -b 128 -w 4 -f ~/tmp/logstash.cfg  1164.86s user 25.72s system 428% cpu 4:37.68 total
pv  3.14s user 19.99s system 8% cpu 4:37.68 total
wc -c  0.44s user 3.58s system 1% cpu 4:37.68 total
```

Gains in both versions via last months fixes reducing function call count and indirection in a few places :)