@original-brownbear 2017-09-28
@andrewvc is there anything I can do here to speed this along? If the main concern here is having Java only tests, I could try and code up the infrastructure for that next week? :)
(I'm just looking for a way of not letting this sit here for another 3 weeks, I already had to rebase it a few times again + one PR merged yesterday did hurt the performance of this already and it was hard to argue on the basis of an unmerged PR)