@andrewvc 2017-09-28
@original-brownbear I will definitely hand off the review if I can't get to it this week.

I think I'm personally not far from LGTM, so I'd rather not do that, which would delay it even further, unless necessary. I promise an answer by tomorrow EOD