@original-brownbear 2017-09-28
@andrewvc alright thanks, I don't think any Java tests are necessary at this point. So yea let's not do them, was just trying a new angle at getting this moving :D