@andrewvc 2017-09-28
After discussing with @ph , and given that this feature is targeted for 6.1.0, not 6.0, what I'd like to propose is to make this an opt-in feature behind a `--experimental-java-execution` flag or similar.

This is amazing work @original-brownbear , but it's too risky to make this the default in a minor without more real-world feedback.

We have me and @ph doing the reviews here, but I think we only really need two reviewers on the new execution code. This means, of course, resurrecting some of the deleted files, but hopefully that isn't too bad? @ph alone is good enough to review any feature flag code.

Once again, great work here, and looking forward to seeing this land in 6.1.0.