@original-brownbear 2017-09-28
@andrewvc 

>  but hopefully that isn't too bad?

This is pretty bad, but what else can we do? :) It's just that:

1. (not so important) We didn't follow this approach from day one and now some internal pipeline APIs were refactored here, making this not really a drop in solution at all now. This will require a bit of effort to get things going side-by-side. 
2. (somewhat important) This requires us to run all tests twice (Ruby specs + integration tests). This def. adds some overhead to the test infrastructure. (also note that some functionality was fixed in the Java exec only, like `XOR` and `NAND` which will require us to ignore some tests for the Ruby exec)
3. (pretty bad) At least from my perspective the plan was to leverage this work to be able to move more logic to Java (especially `Batch` and `Queue` interaction). This will be really tricky to pull off if we have the Ruby exec to support in parallel. In effect, we'd have two diverging codebases to maintain for quite a while.

I can code up a version that sits behind a feature flag, but you have been warned, the overhead from this is pretty big effort and code size wise. But unless I hear anything back on those points, I'll just go ahead and code this up though :) Anything to get some movement here :)