@andrewvc 2017-09-28
@original-brownbear this is definitely up for discussion, but we're in a tricky place. I'll add that I'm not happy with where we are here either, but we're in a tricky spot. We have two choices:

1. Release a major refactor in a minor with major potential for a severe bug
2. Release a feature-flagged version that lets power users test things out

Every large PR we've merged into core has had some degree of nasty bug that wasn't caught until after release. Notable examples: LIR (which didn't impact users because it was only used for a non-existent monitoring UI), Java Event, the persistent queue. 

I want to be clear that this has nothing to do with the quality of your work, which is excellent. This is more a problem of unknown unknowns. Which tests did we not think to write?

The main question here is, where do we lay the burden of dealing with this uncertainty? On our users, or on ourselves as developers? Doing this upgrade without making it an opt-in FF means users don't really get to opt-in to the inherent risk of this major refactor.

I agree that we made a mistake not adopting this strategy earlier. I accept a lot of the responsibility for that. However, it's never too late to make the right decision. I think we were originally planning on being able to target this for 6.0. If we had this same PR open for 6.0.0-beta2 I would say that we are fine to merge as is for two reasons: 1. Users would be able to try it out during the beta process and report problems, 2. The expectations around breakage for a new major are different than they are for a new minor

I'm OK to transition fully over to this code as early as 6.2 if we feel confident that this refactor is solid. I'm also OK with merging this PR without feature flags into master. However, maintaining a non-FF version in master and a FF version in 6.x is probably more work.

WRT some core refactors being delayed here, we have a mountain of other really worthy work, like refactoring the Elasticsearch output. We can cycle back to those things when its more practical.

If you have other ideas for how we can approach this problem, let's discuss them here.